requirements.txt
fastapi
uvicorn[standard]
jinja2
aiosqlite
google-genai
anthropic
openai
youtube-transcript-api>=1.0.0
yt-dlp
httpx
python-dotenv
sse-starlette
markdown
pytest
pytest-asyncio==0.23.8
